requirements.txt
asgiref==3.9.1
Django==5.2.4
Markdown==3.8.2
bleach==6.2.0
pillow==11.3.0
pip==24.3.1
sqlparse==0.5.3
tzdata==2025.2